refactor(loadgen): detach shared NATS health from failure runtime
