+++ tools/loadgen/nats_health.go
@@ -29,7 +29,7 @@ func dialNATSPoolWithMetrics(
 	credsFile string,
 	pool string,
 	metrics *Metrics,
-	observer *failureObserverHealth,
+	observer loadgenNATSObserver,
 ) (*o11ynats.Conn, error) {
 	health := newLoadgenNATSHealth(pool, metrics, nil)
 	if health == nil {
@@ -76,8 +76,12 @@ type loadgenNATSHealth struct {
 	callbackSeen   bool
 	connectedState bool
 	closedState    bool
-	observer       *failureObserverHealth
+	observer       loadgenNATSObserver
 	poolState      *loadgenNATSPoolState
+}
+
+type loadgenNATSObserver interface {
+	Set(bool, time.Time, string)
 }
 
 type loadgenNATSPoolState struct {
@@ -87,7 +91,7 @@ type loadgenNATSPoolState struct {
 	pool           string
 	now            func() time.Time
 	connections    map[*loadgenNATSHealth]bool
-	observer       *failureObserverHealth
+	observer       loadgenNATSObserver
 	disconnectedAt time.Time
 	outageStop     chan struct{}
 	stopped        bool
@@ -247,7 +251,7 @@ func (h *loadgenNATSHealth) updateCurrentOutage() {
 func (s *loadgenNATSPoolState) update(
 	connection *loadgenNATSHealth,
 	connected bool,
-	observer *failureObserverHealth,
+	observer loadgenNATSObserver,
 	reason string,
 ) {
 	if s == nil {
@@ -300,7 +304,7 @@ func (s *loadgenNATSPoolState) update(
 
 func (s *loadgenNATSPoolState) remove(
 	connection *loadgenNATSHealth,
-	observer *failureObserverHealth,
+	observer loadgenNATSObserver,
 	reason string,
 ) {
 	if s == nil {
@@ -429,65 +433,4 @@ func (m *Metrics) stopNATSHealth() {
 	for _, state := range states {
 		state.stop()
 	}
-}
-
-// drainSoakNATS flushes what loadgen has already published, but never past the
-// lease boundary. Waiting is the right default: a pending publish belongs to an
-// operation the ledger has recorded, so dropping it would report data loss
-// loadgen itself caused. Once the lease is at risk that trade reverses, because
-// continuing to emit is the one thing the stop boundary promised not to do --
-// so the connection is closed and the interval is marked inconclusive rather
-// than left looking like a clean run that lost messages.
-// soakDrainableConn is the slice of *nats.Conn the drain needs, so both ways
-// the flush can fail are reachable from a test without a broker.
-type soakDrainableConn interface {
-	Drain() error
-	Close()
-	ClosedHandler() nats.ConnHandler
-	SetClosedHandler(nats.ConnHandler)
-}
-
-func drainSoakNATS(
-	nc soakDrainableConn,
-	budget time.Duration,
-	invalidate func(string),
-) {
-	if nc == nil {
-		return
-	}
-	if budget <= 0 {
-		// Ordinary shutdown, unchanged: start the drain and let the process
-		// finish on its own schedule. No lease is at stake, so a refusal here
-		// says nothing about the system under test.
-		if err := nc.Drain(); err != nil {
-			slog.Error("drain Cassandra soak NATS connection", "error", err)
-		}
-		return
-	}
-	closed := make(chan struct{})
-	previous := nc.ClosedHandler()
-	nc.SetClosedHandler(func(conn *nats.Conn) {
-		if previous != nil {
-			previous(conn)
-		}
-		close(closed)
-	})
-	// Drain refuses outright on a closed or already-draining connection, which
-	// leaves the pending publishes in the same unknown state as a drain that
-	// ran out of budget. Both are the same fact about the evidence, so they get
-	// the same answer.
-	err := nc.Drain()
-	if err == nil && waitSoakDrain(closed, budget) {
-		return
-	}
-	slog.Error(
-		"abandoned the soak NATS drain at the lease boundary",
-		"budget", budget,
-		"error", err,
-		"consequence", "unflushed publishes are dropped; the interval is inconclusive",
-	)
-	if invalidate != nil {
-		invalidate(invalidReasonLeaseAbort)
-	}
-	nc.Close()
 }
